"Index.pkl file has been updated with a new version, binary files differ."
This commit message is based on the provided code diff, which shows that there has been a modification to the file 'index.pk' in the 'faiss\_index' directory. The 'index.pkl' file is a binary file, and the diff shows that the new version of the file (0a66a5f) is different from the old version (9fd6b68). Therefore, the commit message should reflect this change by stating that the 'index.pkl' file has been updated with a new version, and that the binary files differ.